refactor: let method to repository_impl.go
